+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # Changelog
 
+## [Unreleased]
+
 ## [0.0.2]
 - SOMETHING NEW
 
 ## [0.0.1]
 - BREAKING CHANGE: API class names are CamelCase. This replaces the  "lower  case version" naming.
+
+[Unreleased]: https://github.com/bg-omar/consolelogger/compare/v0.0.2...HEAD
 
 [0.0.2]: https://github.com/bg-omar/consolelogger/compare/v0.0.1...v0.0.2